+++ src/main/java/com/KoiHealthService/Koi/demo/dto/response/AppointmentTypeResponse.java
@@ -14,5 +14,5 @@ import java.math.BigDecimal;
 public class AppointmentTypeResponse {
     String appointmentTypeId;
     String appointmentService;
-    BigDecimal price;
+    Long price;
 }
